requirements.txt
openai>=1.3.5
requests
beautifulsoup4
psycopg2-binary
PyMuPDF
python-dotenv
pytest
fpdf
